refactor: unify storage barrel imports and tidy hook.ts
